Add rate control functions to the URL location updating
Updating the URL too often was throwing some errors. Have added a setting to allow the delay to be set as to how often this can occur in milliseconds. Setting the delay to zero with go back to previous functionality of live updating, and setting to -1 disables completely and after domain connect it remove the location from the url.